Update `RecyclerView` appearance and logic
