+++ app/src/main/java/com/senex/androidlab1/views/activities/MainActivity.kt
@@ -41,10 +41,5 @@ class MainActivity : AppCompatActivity() {
             //*generateUsers(20).toTypedArray()
         )
     }
-
-    override fun onSupportNavigateUp(): Boolean {
-        onBackPressed()
-        return true
-    }
 }
