+++ app/src/main/java/com/senex/androidlab1/views/fragments/InteractiveListFragment.kt
@@ -7,12 +7,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.senex.androidlab1.adapters.UserRecyclerAdapter
+import com.senex.androidlab1.adapters.ListRecyclerAdapter
 import com.senex.androidlab1.database.AppDatabaseMain
 import com.senex.androidlab1.databinding.FragmentInteractiveListBinding
-import com.senex.androidlab1.databinding.FragmentListBinding
 import com.senex.androidlab1.utils.MarginItemDecoration
-
 
 class InteractiveListFragment : Fragment() {
     private var _binding: FragmentInteractiveListBinding? = null
@@ -36,11 +34,9 @@ class InteractiveListFragment : Fragment() {
             requireContext()
         )
 
-        adapter = UserRecyclerAdapter(
+        adapter = ListRecyclerAdapter(
             AppDatabaseMain.database.userDao().getAll()
-        ) {
-            // Setup onclick listener
-        }
+        )
 
         addItemDecoration(
             MarginItemDecoration(20)
