+++ app/src/main/res/layout/interactive_list_item.xml
